+++ R/marker_genomatrix.R
@@ -19,13 +19,11 @@
 #' @seealso [getAlleles()]
 #'
 #' @examples
-#' x = nuclearPed(1)
-#' m1 = marker(x, `2` = "1/2", alleles = 1:2, name = "m1")
-#' m2 = marker(x, `3` = "2/2", alleles = 1:2, name = "m2")
-#' x = setMarkers(x, list(m1, m2))
+#' x = nuclearPed() |>
+#'   addMarker(`2` = "1/2", name = "m1") |>
+#'   addMarker(`3` = "a/a", name = "m2")
 #'
 #' getGenotypes(x)
-#'
 #'
 #' ### A list of pedigrees
 #'
@@ -46,8 +44,8 @@ getGenotypes = function(x, ids = NULL, markers = NULL, sep = "/", missing = "-")
   # Check `ids` argument
   if(!is.null(ids)) {
     ids = as.character(ids)
-    if(!all(ids %in% unlist(labels(x))))
-      stop2("Unknown ID label: ", setdiff(ids, unlist(labels(x))))
+    if(!all(ids %in% labels(x)))
+      stop2("Unknown ID label: ", setdiff(ids, labels(x)))
   }
 
   if(is.pedList(x)) {