+++ R/ped_labels.R
@@ -164,8 +164,10 @@ relabel = function(x, new = "asPlot", old = labels(x), reorder = FALSE,
   x
 }
 
-#' @param object A `ped` object
-#' @param ... Not used
+#' @param object A `ped` object.
+#' @param ... Not used.
+#' @param unlist A logical; if TRUE (default), the output is unlisted to a
+#'   single character vector.
 #'
 #' @rdname relabel
 #' @export
@@ -175,9 +177,14 @@ labels.ped = function(object, ...) {
 
 #' @rdname relabel
 #' @export
-labels.list = function(object, ...) {
-  if(is.pedList(object))
-    lapply(object, labels.ped)
-  else
-    labels.default(object)
+labels.list = function(object, ..., unlist = TRUE) {
+  if(is.pedList(object)) {
+    labs = lapply(object, labels.ped)
+    if(unlist)
+      labs = unlist(labs, recursive = FALSE, use.names = FALSE)
+    return(labs)
+  }
+
+  # else
+  labels.default(object)
 }