+++ R/ped_plot.R
@@ -13,7 +13,7 @@
 #' duplications.
 #'
 #' If `x` is a list of `ped` objects these are plotted next to each other,
-#' vertically centered in the plot window. For finer control, and possibly
+#' vertically centred in the plot window. For finer control, and possibly
 #' nested lists of pedigrees, use [plotPedList()].
 #'
 #' @param x A [ped()] object or a list of such.
@@ -54,8 +54,7 @@
 #' # Various annotations
 #' plot(x, hatched = "boy", starred = "fa", deceased = "mo", title = "Fam 1")
 #'
-#' # Medical pedigree
-#' plot(x, aff = "boy", carrier = "mo")
+#' #----- ID labels -----
 #'
 #' # Label only some members
 #' plot(x, labs = c("fa", "mo"))
@@ -66,13 +65,23 @@
 #' # Rename some individuals
 #' plot(x, labs = c(FATHER = "fa", "boy"))
 #'
+#' # By default, long names are folded to width ~12 characters
+#' plot(x, labs = c("Very long father's name" = "fa"), margin = 2)
+#'
+#' # Folding width may be adjusted ...
+#' plot(x, labs = c("Very long father's name" = "fa"), foldLabs = 6)
+#'
+#' # ... or switched off (requires larger margin!)
+#' plot(x, labs = c("Very long father's name" = "fa"), foldLabs = FALSE)
+#'
 #' # By default, labels are trimmed for initial/trailing line breaks ...
 #' plot(x, labs = c("\nFA" = "fa"))
 #'
 #' # ... but this can be overridden
 #' plot(x, labs = c("\nFA" = "fa"), trimLabs = FALSE)
 #'
-#' # Colours
+#' #----- Colours -----
+#'
 #' plot(x, col = c(fa = "red"), fill = c(mo = "green", boy = "blue"))
 #'
 #' # Non-black hatch colours are specified with the `fill` argument
@@ -81,34 +90,61 @@
 #' # Use functions to specify colours
 #' plot(x, fill = list(red = leaves, blue = ancestors(x, "boy")))
 #'
-#' # Line type and width
-#' plot(x, lty = 2, lwd = 3, cex = 2)
+#' #----- Symbol line types and widths -----
 #'
-#' # Detailed line type and width
+#' # Dotted, thick symbols
+#' plot(x, lty = 3, lwd = 4, cex = 2)
+#'
+#' # Detailed specification of line types and width
 #' plot(x, lty = list(dashed = founders), lwd = c(boy = 4))
 #'
-#' # Include genotypes
-#' x = addMarker(x, fa = "1/1", boy = "1/2", name = "SNP")
+#' #----- Genotypes -----
+#'
+#' x = nuclearPed(father = "fa", mother = "mo", child = "boy") |>
+#'   addMarker(fa = "1/1", boy = "1/2", name = "SNP") |>
+#'   addMarker(boy = "a/b")
+#'
+#' # Show genotypes for first marker
 #' plot(x, marker = 1)
+#'
+#' # Show empty genotypes for untyped individuas
+#' plot(x, marker = 1, showEmpty = TRUE)
 #'
 #' # Markers can also be called by name
 #' plot(x, marker = "SNP")
 #'
+#' # Multiple markers
+#' plot(x, marker = 1:2)
+#'
+#' #----- Further text annotation -----
+#'
+#' # Founder inbreeding is shown by default
+#' xinb = x |> setFounderInbreeding("mo", value = 0.1)
+#' plot(xinb)
+#'
+#' # ... but can be suppressed
+#' plot(xinb, fouInb = NULL)
+#'
+#' # Text can be placed around and inside symbols
+#' plot(x, textAnnot = list(topright = 1:3, inside = LETTERS[1:3]))
+#'
+#' # Use lists to add further options; see `?text()`
+#' plot(x, margin = 2, textAnnot = list(
+#'   topright = list(1:3, cex = 0.8, col = 2, font = 2, offset = 0.1),
+#'   left = list(c(boy = "comment"), cex = 2, col = 4, offset = 2, srt = 20)))
+#'
+#' # Exhaustive list of annotation positions
+#' plot(singleton(1), cex = 3, textAnnot = list(top="top", left="left",
+#'   right="right", bottom="bottom", topleft="topleft", topright="topright",
+#'   bottomleft="bottomleft", bottomright="bottomright", inside="inside"))
+#'
+#' #----- Special pedigrees -----
+#'
 #' # Plot as DAG (directed acyclic graph)
 #' plot(x, arrows = TRUE, title = "DAG")
 #'
-#' # Founder inbreeding is shown by default
-#' founderInbreeding(x, "mo") = 0.1
-#' plot(x)
-#'
-#' # ... but can be suppressed
-#' plot(x, fouInb = NULL)
-#'
-#' # Other text above and inside symbols
-#' plot(x, textAbove = letters[1:3], textInside = LETTERS[1:3])
-#'
-#' # Plotting lists of pedigrees
-#' plot(list(singleton(1), nuclearPed(1), linearPed(2)))
+#' # Medical pedigree
+#' plot(x, aff = "boy", carrier = "mo")
 #'
 #' # Twins
 #' x = nuclearPed(children = c("tw1", "tw2", "tw3"))
@@ -128,6 +164,9 @@
 #'
 #' # Straight legs
 #' plot(quadHalfFirstCousins(), align = c(0,0))
+#'
+#' # Lists of multiple pedigree
+#' plot(list(singleton(1), nuclearPed(1), linearPed(2)))
 #'
 #' # Use of `drawPed()`
 #' dat = plot(nuclearPed(), draw = FALSE)
@@ -418,7 +457,10 @@ plotPedList = function(plots, widths = NULL, groups = NULL, titles = NULL,
     else if (is.pedList(p))
       newpeds = lapply(p, list)
     else { # if list of ped with plot arguments
-      if (!is.ped(p[[1]]))
+      p1 = p[[1]]
+      if(inherits(p[[1]], "pedList"))
+        class(p[[1]]) = "list"
+      else if (!is.ped(p[[1]]))
         stop2("First element must be a `ped` object", p[[1]])
       newpeds = list(p)
     }
@@ -652,15 +694,26 @@ plot.list = function(x, ...) {
   annotList = lapply(x, function(y) .pedAnnotation(y, ...))
   annot1 = annotList[[1]]
 
+  # Merge vectors
   mrg = function(a, def) {
     vecs = lapply(1:L, function(i) annotList[[i]][[a]] %||% rep(def, nInd[i]))
     unlist(vecs)
+  }
+
+  # Merge lists
+  mrgTxtGen = function() {
+    res1 = annotList[[1]]$textAnnot
+    for(nm in names(res1))
+      res1[[nm]][[1]] = unlist(lapply(1:L, function(i) annotList[[i]]$textAnnot[[nm]][[1]]))
+
+    res1
   }
 
   annot = list(title = annot1$title,
                textUnder = mrg("textUnder", ""),
                textAbove = mrg("textAbove", ""),
                textInside = mrg("textInside", ""),
+               textAnnot = mrgTxtGen(),
                colvec = mrg("colvec", 1),
                densvec = mrg("densvec", 0),
                fillvec =  mrg("fillvec", NA),