+++ R/ped_utils.R
@@ -351,7 +351,31 @@ hasNumLabs = function(x) {
   isTRUE(all(labs == numlabs))
 }
 
+# New implementation of `hasNumLabs`; takes either a character or ped(list)
+.isIntegral = function(labs) {
+  if(!is.character(labs))
+    labs = labels(labs)
 
+  intlabs = suppressWarnings(as.character(as.integer(labs)))
+  isTRUE(all(labs == intlabs))
+}
+
+# Utility for creating new labels
+generateLabs = function(x, n = 1, avoid = NULL, prefix = "") {
+  labs = if(is.character(x)) x else labels(x)
+
+  # Simple strategy (no need to be clever about this...)
+  # Generate enough candidates and drop the taken ones
+  taken = c(labs, avoid)
+  cand = as.character(seq_len(length(taken) + n))
+
+  if(nzchar(prefix))
+    cand = paste0(prefix, cand)
+
+  .mysetdiff(cand, taken)[seq_len(n)]
+}
+
+# TODO: Delete - no longer used. Replaced by generateLabs
 # Utility function for generating numbered "NN" labels.
 # Returns "NN_i" where i increments largest j occurring as NN_j, NN.j or NN-j in input.
 nextNN = function(labs) { # labs a character vector
@@ -398,10 +422,10 @@ getComponent = function(x, ids, checkUnique = FALSE, errorIfUnknown = FALSE) {
   }
 
   # List labels of each component
-  labList = labels(x)
+  labList = labels(x, unlist = FALSE)
 
   # A single vector with all labels
-  labVec = unlist(labList)
+  labVec = unlist(labList, recursive = FALSE, use.names = FALSE)
 
   # Check for duplicates if indicated
   if(checkUnique) {